Better Formatting
Formatted all HTML, CSS, and JS. A lot of the source seemed to be fairly dated or cryptic, so I tidied up and modernized it to be more up to date with today's coding practices. I didn't make any actual changes to the site itself from the user's perspective, only from the developer's point of view.