Finally Updated the README
Finally got around to updating the README! It was so old, from back before I called the project Minefork.
I was going to add more info to it for the longest time, but never got around to doing anything fancy with it. Eventually I want to have more written up about the project, and a few pictures would be cool too.